requirements.txt
matplotlib==2.2.2
requests==2.18.4
PyExecJS==1.5.1
